--- LICENSE
@@ -1,3 +1,0 @@
-This project is developed as part of a school assignment.  
-Copyright © 2024 @Loulouplou and @KarolannMauger.  
-All rights reserved. Redistribution, modification, and commercial use are strictly prohibited.
