--- README.md
@@ -1,18 +1,0 @@
-# **saPin**  
-
-A Raspberry Pi project for Christmas.  
-
-## Getting Started  
-The project is under development. More details will be added soon.  
-
-## Installation  
-Instructions will be available once the project is ready.
-
-## License  
-This project is developed as part of the course *PROGRAMMATION DE PLATEFORMES EMBARQUÉES | 420-314-MV*.  
-Copyright © [2024] [@Loulouplou](https://www.github.com/Loulouplou) and [@KarolannMauger](https://www.github.com/KarolannMauger).  
-All rights reserved. Redistribution, modification, and commercial use are strictly prohibited.  
-
----
-
-🎅 *Happy holidays from the saPin team!*
